http_bench_bin_ops works
